+++ crates/holochain/tests/tests/warrants.rs
@@ -43,7 +43,7 @@ async fn warranted_agent_is_blocked() {
     let (bob_conductor, bob_cell) = conductors_and_cells.remove(0);
 
     // Let all agents sync.
-    await_consistency(10, [&alice_cell, &bob_cell])
+    await_consistency(15, [&alice_cell, &bob_cell])
         .await
         .unwrap();
 
@@ -56,7 +56,7 @@ async fn warranted_agent_is_blocked() {
         )
         .await;
 
-    await_consistency(10, [&alice_cell, &bob_cell])
+    await_consistency(15, [&alice_cell, &bob_cell])
         .await
         .unwrap();
 
@@ -121,7 +121,7 @@ async fn warrant_is_gossiped() {
     let (_bob_conductor, bob_cell) = conductors_and_cells.remove(0);
     let (carol_conductor, carol_cell) = conductors_and_cells.remove(0);
 
-    await_consistency(10, [&alice_cell, &bob_cell, &carol_cell])
+    await_consistency(15, [&alice_cell, &bob_cell, &carol_cell])
         .await
         .unwrap();
 
@@ -238,7 +238,7 @@ async fn author_of_invalid_warrant_is_blocked() {
         .await;
 
     // Wait for Alice and Bob to sync.
-    await_consistency(10, [&alice, &bob]).await.unwrap();
+    await_consistency(15, [&alice, &bob]).await.unwrap();
 
     let alice_authored_db = conductors[0]
         .get_spaces()
@@ -288,7 +288,7 @@ async fn author_of_invalid_warrant_is_blocked() {
         });
 
     // Wait for Alice and Bob to sync so that Alice receives the warrant.
-    await_consistency(10, [&alice, &bob]).await.unwrap();
+    await_consistency(15, [&alice, &bob]).await.unwrap();
 
     tokio::time::timeout(std::time::Duration::from_secs(30), async {
         loop {
@@ -368,7 +368,7 @@ mod zero_arc {
         let (bob_conductor, bob_cell) = conductors_and_cells.remove(0);
         let (carol_conductor, carol_cell) = conductors_and_cells.remove(0);
 
-        await_consistency(10, [&alice_cell, &bob_cell])
+        await_consistency(15, [&alice_cell, &bob_cell])
             .await
             .unwrap();
         bob_conductor
@@ -387,7 +387,7 @@ mod zero_arc {
             )
             .await;
 
-        await_consistency(10, [&alice_cell, &bob_cell])
+        await_consistency(15, [&alice_cell, &bob_cell])
             .await
             .unwrap();
 
@@ -433,7 +433,7 @@ mod zero_arc {
         let (bob_conductor, bob_cell) = conductors_and_cells.remove(0);
         let (carol_conductor, carol_cell) = conductors_and_cells.remove(0);
 
-        await_consistency(10, [&alice_cell, &bob_cell])
+        await_consistency(15, [&alice_cell, &bob_cell])
             .await
             .unwrap();
 
@@ -454,7 +454,7 @@ mod zero_arc {
             )
             .await;
 
-        await_consistency(10, [&alice_cell, &bob_cell])
+        await_consistency(15, [&alice_cell, &bob_cell])
             .await
             .unwrap();
 
@@ -517,7 +517,7 @@ mod zero_arc {
         let (bob_conductor, bob_cell) = conductors_and_cells.remove(0);
         let (carol_conductor, carol_cell) = conductors_and_cells.remove(0);
 
-        await_consistency(10, [&alice_cell, &bob_cell])
+        await_consistency(15, [&alice_cell, &bob_cell])
             .await
             .unwrap();
 
@@ -538,7 +538,7 @@ mod zero_arc {
             )
             .await;
 
-        await_consistency(10, [&alice_cell, &bob_cell])
+        await_consistency(20, [&alice_cell, &bob_cell])
             .await
             .unwrap();