+++ crates/holochain_p2p/src/spawn/actor.rs
@@ -892,12 +892,27 @@ impl HolochainP2pActor {
 
         #[cfg(feature = "test_utils")]
         {
+            #[cfg(any(
+                feature = "transport-tx5-datachannel-vendored",
+                feature = "transport-tx5-backend-libdatachannel",
+                feature = "transport-tx5-backend-go-pion",
+            ))]
             builder
                 .config
                 .set_module_config(&kitsune2_transport_tx5::Tx5TransportModConfig {
                     tx5_transport: kitsune2_transport_tx5::Tx5TransportConfig {
                         signal_allow_plain_text: true,
                         timeout_s: 20,
+                        webrtc_connect_timeout_s: 15,
+                        ..Default::default()
+                    },
+                })?;
+            #[cfg(feature = "transport-iroh")]
+            builder
+                .config
+                .set_module_config(&kitsune2_transport_iroh::IrohTransportModConfig {
+                    iroh_transport: kitsune2_transport_iroh::IrohTransportConfig {
+                        relay_allow_plain_text: true,
                         ..Default::default()
                     },
                 })?;
@@ -954,6 +969,32 @@ impl HolochainP2pActor {
             pruning_task_abort_handle,
             request_timeout: config.request_timeout,
         }))
+    }
+
+    /// Extract Kitsune2 [`Config`] from a [`serde_json::Value`].
+    fn kitsune2_params_from_value(value: serde_json::Value) -> HolochainP2pResult<Config> {
+        let config = Config::default();
+        // get `core_bootstrap` from config
+        if let Ok(core_bootstrap_config) = serde_json::from_value::<
+            kitsune2_core::factories::CoreBootstrapModConfig,
+        >(value.clone())
+        {
+            config.set_module_config(&core_bootstrap_config)?;
+        }
+
+        // get `tx5_transport` from config
+        #[cfg(any(
+            feature = "transport-tx5-datachannel-vendored",
+            feature = "transport-tx5-backend-libdatachannel",
+            feature = "transport-tx5-backend-go-pion",
+        ))]
+        if let Ok(tx5_transport_config) =
+            serde_json::from_value::<kitsune2_transport_tx5::Tx5TransportModConfig>(value)
+        {
+            config.set_module_config(&tx5_transport_config)?;
+        }
+
+        Ok(config)
     }
 
     // Prunes expired URLs at an interval and checks the peer store for agent infos of unresponsive
@@ -2222,6 +2263,12 @@ impl actor::HcP2p for HolochainP2pActor {
 #[cfg(test)]
 mod tests {
     use super::*;
+    #[cfg(any(
+        feature = "transport-tx5-backend-libdatachannel",
+        feature = "transport-tx5-backend-go-pion",
+        feature = "transport-tx5-datachannel-vendored"
+    ))]
+    use crate::actor::HcP2p;
 
     #[tokio::test(flavor = "multi_thread")]
     async fn correct_id_loc_calc() {