+++ crates/holochain_p2p/src/types/actor.rs
@@ -184,7 +184,7 @@ pub trait HcP2p: 'static + Send + Sync + std::fmt::Debug {
                         })
                     }
                 },
-                Some(10_000),
+                Some(20_000),
                 None,
             )
             .await