+++ crates/holochain_p2p/tests/tests/node_messaging.rs
@@ -981,6 +981,8 @@ async fn test_must_get_agent_activity_with_unresponsive_agents() {
 
 #[tokio::test(flavor = "multi_thread")]
 async fn test_validation_receipts() {
+    test_run();
+
     let dna_hash = DnaHash::from_raw_36(vec![0; 36]);
     let handler = Arc::new(Handler::default());
 
@@ -1219,6 +1221,11 @@ async fn spawn_test(
                 let conductor_db = conductor_db.clone();
                 Box::pin(async move { conductor_db })
             }),
+            #[cfg(any(
+                feature = "transport-tx5-datachannel-vendored",
+                feature = "transport-tx5-backend-libdatachannel",
+                feature = "transport-tx5-backend-go-pion"
+            ))]
             network_config: Some(serde_json::json!({
                 "coreBootstrap": {
                     "serverUrl": format!("http://{bootstrap_addr}"),
@@ -1230,7 +1237,17 @@ async fn spawn_test(
                     "webrtcConnectTimeoutS": 25,
                 }
             })),
-            request_timeout: Duration::from_secs(3),
+            #[cfg(feature = "transport-iroh")]
+            network_config: Some(serde_json::json!({
+                "coreBootstrap": {
+                    "serverUrl": format!("http://{bootstrap_addr}"),
+                },
+                "irohTransport": {
+                    "relayUrl": format!("http://{bootstrap_addr}"),
+                    "relayAllowPlainText": true,
+                }
+            })),
+            request_timeout: Duration::from_secs(10),
             ..Default::default()
         },
         lair_client.clone(),
@@ -1240,7 +1257,25 @@ async fn spawn_test(
 
     hc.register_handler(handler).await.unwrap();
 
-    hc.join(dna_hash, agent.clone(), None).await.unwrap();
+    hc.join(dna_hash.clone(), agent.clone(), None)
+        .await
+        .unwrap();
+
+    // Wait for the endpoint to have a current URL.
+    retry_fn_until_timeout(
+        || async {
+            hc.test_kitsune()
+                .space(dna_hash.to_k2_space(), None)
+                .await
+                .unwrap()
+                .current_url()
+                .is_some()
+        },
+        Some(20_000),
+        None,
+    )
+    .await
+    .unwrap();
 
     (agent, hc, lair_client)
 }